requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
motor>=3.3.0
redis>=5.0.0
pandas>=2.0.0
numpy>=1.24.0
ta-lib>=0.4.0
requests>=2.31.0
python-dotenv>=1.0.0
google-genai
beautifulsoup4>=4.12.0
lxml>=4.9.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.24.0
yfinance>=0.2.0
